[#1702] Create iceoryx2 testing library
