+++ iceoryx2/conformance-tests/Cargo.toml
@@ -20,6 +20,7 @@ std = [
   "iceoryx2-bb-elementary/std",
   "iceoryx2-bb-posix/std",
   "iceoryx2-bb-testing/std",
+  "iceoryx2-testing/std",
   "iceoryx2-bb-testing-macros/std",
   "iceoryx2-cal/std",
   "iceoryx2-log/std",
@@ -33,6 +34,7 @@ harness = false
 
 [dependencies]
 iceoryx2 = { workspace = true }
+iceoryx2-testing = { workspace = true }
 iceoryx2-bb-concurrency = { workspace = true }
 iceoryx2-bb-container = { workspace = true }
 iceoryx2-bb-derive-macros = { workspace = true }
