+++ iceoryx2/testing/src/lib.rs
@@ -1,0 +1,104 @@
+// Copyright (c) 2026 Contributors to the Eclipse Foundation
+//
+// See the NOTICE file(s) distributed with this work for additional
+// information regarding copyright ownership.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Apache Software License 2.0 which is available at
+// https://www.apache.org/licenses/LICENSE-2.0, or the MIT license
+// which is available at https://opensource.org/licenses/MIT.
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+pub use iceoryx2::testing::*;
+
+use std::marker::PhantomData;
+use std::time::Duration;
+
+use iceoryx2::prelude::*;
+use iceoryx2_bb_elementary_traits::testing::abandonable::Abandonable;
+use iceoryx2_bb_testing::watchdog::Watchdog;
+
+pub struct Test<Service: iceoryx2::service::Service> {
+    config: Config,
+    _watchdog: Watchdog,
+    _data: PhantomData<Service>,
+}
+
+impl<Service: iceoryx2::service::Service> Drop for Test<Service> {
+    fn drop(&mut self) {
+        Self::cleanup_dead_nodes(&self.config);
+        unsafe { remove_global_mgmt_segment::<Service>(&self.config).unwrap() };
+    }
+}
+
+impl<Service: iceoryx2::service::Service> Default for Test<Service> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl<Service: iceoryx2::service::Service> Test<Service> {
+    pub fn new() -> Self {
+        let _watchdog = Watchdog::new();
+        let mut config = generate_isolated_config();
+        config.global.node.cleanup_dead_nodes_on_creation = false;
+        config.global.node.cleanup_dead_nodes_on_destruction = false;
+        config.global.service.cleanup_dead_nodes_on_open = false;
+
+        Self {
+            config,
+            _watchdog,
+            _data: PhantomData,
+        }
+    }
+
+    pub fn config(&self) -> &Config {
+        &self.config
+    }
+
+    pub fn config_mut(&mut self) -> &mut Config {
+        &mut self.config
+    }
+
+    pub fn abandon_contents<T: Abandonable>(&self, mut contents: Vec<T>) {
+        while let Some(element) = contents.pop() {
+            T::abandon(element);
+        }
+    }
+
+    pub fn create_node(&self) -> Node<Service> {
+        NodeBuilder::new()
+            .config(self.config())
+            .create::<Service>()
+            .unwrap()
+    }
+
+    pub fn number_of_nodes(&self) -> usize {
+        self.list_nodes().len()
+    }
+
+    pub fn list_nodes(&self) -> Vec<NodeState<Service>> {
+        let mut node_list = vec![];
+        Node::<Service>::list(self.config(), |node_state| {
+            node_list.push(node_state);
+            CallbackProgression::Continue
+        })
+        .unwrap();
+
+        node_list
+    }
+
+    pub fn cleanup_dead_nodes(config: &Config) {
+        Node::<Service>::list(config, |node_state| {
+            if let NodeState::Dead(state) = node_state {
+                state
+                    .blocking_remove_stale_resources(Duration::MAX)
+                    .unwrap();
+            }
+
+            CallbackProgression::Continue
+        })
+        .unwrap();
+    }
+}
